+++ extension/src/content-script/osu/pages/topics.ts
@@ -1,7 +1,8 @@
 // https://osu.ppy.sh/community/forums/topics/1686524?n=3
 
-import { addFlagUser } from "@src/content-script/osu/flagHtml";
-import { nextAbortControllerSignal } from "./content";
+import { addFlagUsers as addFlagUsers } from "@src/content-script/osu/flagHtml";
+import { TFlagItems } from "@src/utils/html";
+import { nextAbortControllerSignal } from "@src/utils/fetchUtils";
 
 export const updateFlagsTopics = async () => {
     const url = location.href;
@@ -9,17 +10,18 @@ export const updateFlagsTopics = async () => {
     const signal = nextAbortControllerSignal();
     const posts = document.querySelectorAll(".forum-post-info");
 
+    const flagItems: TFlagItems = [];
+
     for (const item of posts) {
-        if (signal.aborted) {
-            return;
-        }
         const playerNameElement = item.querySelector(".forum-post-info__row--username") as HTMLElement;
         const playerId = playerNameElement.getAttribute("data-user-id")!;
-        addFlagUser(item as HTMLElement, playerId, {
-            addDiv: false,
-            addMargin: false,
-            addSuperParentClone: true,
-            signal: signal,
-        });
+        flagItems.push({ id: playerId, item: item as HTMLElement });
     }
+    
+    await addFlagUsers(flagItems, {
+        addDiv: false,
+        addMargin: false,
+        addSuperParentClone: true,
+        signal: signal,
+    });
 };