+++ extension/src/utils/fetchUtils.ts
@@ -20,18 +20,21 @@ export interface IfetchResponse<T> {
 }
 
 export type fetchOptions = {
-    preserve?: boolean;
     signal?: AbortSignal;
-};
+}
 
-const genExpireDate = (expireTime: number) => Date.now() + expireTime;
+export type fetchCacheOptions = {
+    preserve?: boolean;
+} & fetchOptions;
+
+export const genExpireDate = (expireTime: number) => Date.now() + expireTime;
 
 const pendingRequests: { [key: string]: Promise<object> } = {};
 
 const fetchAndSaveInCache = async (
     url: string,
     expireTime: number,
-    options: fetchOptions,
+    options: fetchCacheOptions,
 ): Promise<IfetchResponse<object>> => {
     const { preserve, signal } = options;
     return fetch(url, { signal: signal })
@@ -46,11 +49,14 @@ const fetchAndSaveInCache = async (
                 };
             const cachedResponse: IfetchResponse<object> = {};
             cachedResponse["data"] = jsonResponse;
-            cachedResponse[expireHeader] = genExpireDate(expireTime);
-            if (preserve) {
-                cachedResponse["preserve"] = true;
+            if(expireTime > 0){
+                cachedResponse[expireHeader] = genExpireDate(expireTime);
+                if (preserve) {
+                    cachedResponse["preserve"] = true;
+                }
+                saveInCache(url, cachedResponse);
             }
-            saveInCache(url, cachedResponse);
+
             return cachedResponse;
         })
         .catch((err) => {
@@ -72,7 +78,7 @@ const fetchAndSaveInCache = async (
 export const fetchWithCache = async (
     url: string,
     expireTime: number,
-    options: fetchOptions = {},
+    options: fetchCacheOptions = {},
 ): Promise<IfetchResponse<object>> => {
     if (pendingRequests[url] !== undefined) {
         return pendingRequests[url];
@@ -95,16 +101,28 @@ export const fetchWithCache = async (
     }
 };
 
+export const fetchWithoutCache = async (url: string, options: fetchOptions = {}) : Promise<IfetchResponse<object>> => {
+    if (pendingRequests[url] !== undefined) {
+        return pendingRequests[url];
+    }
+    const fetchPromise = fetchAndSaveInCache(url, 0, options);
+    pendingRequests[url] = fetchPromise;
+    const result = await fetchPromise;
+    delete pendingRequests[url];
+    return result;
+}
+
 export const unknownUserError = "unknown_user";
 const cannotFetchError = "cannot_fetch";
-const noData = "no_data";
+export const noData = "no_data";
 export const noId = "no_id";
 export const noMode = "no_mode";
 
 export const fetchErrorToText = (response: IfetchResponse<object> | undefined) => {
-    if (!response?.error?.code) return "";
-    const error = response.error;
-    switch (error.code) {
+    const errorCode = response?.error?.code ?? response?.error;
+    if(!errorCode) return "";
+    const error = response?.error!;
+    switch (errorCode) {
         case unknownUserError:
             return "Unknown user " + error.userId;
         case cannotFetchError:
@@ -112,7 +130,7 @@ export const fetchErrorToText = (response: IfetchResponse<object> | undefined) =
         case noData:
             return "No data for " + error.url;
         case noId:
-            return "No player found with ID " + error.userId;
+            return "No player ID found";
         case noMode:
             return "No mode found when paring url " + error.mode;
         default:
@@ -146,3 +164,15 @@ export const fetchWithMinimumWaitTime = async <T>(
             return result as IfetchResponse<T>;
         });
 };
+
+export let currentAbortController = new AbortController();
+
+
+export const currentSignal = () => currentAbortController.signal;
+export const nextAbortControllerSignal = () => {
+    currentAbortController.abort();
+    currentAbortController = new AbortController();
+
+    return currentAbortController.signal;
+};
+