+++ extension/src/utils/osuWorld.ts
@@ -1,3 +1,4 @@
+import { loadFromCache, loadMultipleUrlsFromCache, saveInCache } from "./cache";
 import {
     IfetchResponse,
     noId,
@@ -5,6 +6,11 @@ import {
     fetchWithMinimumWaitTime,
     IFetchError,
     fetchErrorToText,
+    fetchWithoutCache,
+    genExpireDate,
+    expireHeader,
+    unknownUserError,
+    noData,
 } from "./fetchUtils";
 
 export interface IosuWorldIdSuccess {
@@ -28,30 +34,156 @@ export interface IosuWorldRegionalRankingSuccess {
     top: [IosuWorldRegionalPlayerData];
 }
 
-export type osuWorldIdData = IosuWorldIdSuccess | IFetchError;
+export type TosuWorldIdData = IosuWorldIdSuccess | IFetchError;
+export type TosuWorldIdsData = IosuWorldIdSuccess[] | IFetchError;
 
-const userDataExpireTime = 43200000; //12 hours
+const osuWorldApiBase = "https://osuworld.octo.moe/api/";
+
+const userDataExpireTime = 3600000; //60 minutes
 
 export const osuWorldUser = async (
     id: string,
     signal: AbortSignal | undefined,
-): Promise<IfetchResponse<osuWorldIdData>> => {
+): Promise<IfetchResponse<TosuWorldIdData>> => {
     if (!id) {
         console.log("id is null");
-        return { error: { code: noId, userId: id } };
+        return { error: { code: noId } };
     }
 
-    const url = "https://osuworld.octo.moe/api/users/" + id;
+    const url = osuWorldApiBase + "users/" + id;
 
     const dataPromise = fetchWithCache(url, userDataExpireTime, { signal: signal }) as Promise<
-        IfetchResponse<osuWorldIdData>
+        IfetchResponse<TosuWorldIdData>
     >;
-    return fetchWithMinimumWaitTime<osuWorldIdData>(dataPromise, 200);
+    return fetchWithMinimumWaitTime<TosuWorldIdData>(dataPromise, 200);
 };
+
+
+export const osuWorldUsers = async (
+    ids: string[],
+    signal: AbortSignal | undefined,
+): Promise<IfetchResponse<TosuWorldIdsData>> => {
+    if (!ids || ids.length === 0) {
+        console.log("id is null");
+        return { error: { code: noId } };
+    }
+
+    ids = [...new Set(ids)];
+
+    // Max 50 ids per request
+    const promises = [];
+    const urls = [];
+
+    for (let i = 0; i < ids.length; i += 50) {
+        const slicedIds = ids.slice(i, i + 50);
+        const url = osuWorldApiBase + "subdiv/users?ids=" + slicedIds.join(",");
+        urls.push(url);
+        const promise = loadFromCache(url).then((cacheResponse) => {
+            const expireTime = cacheResponse?.[expireHeader];
+            if(!expireTime || expireTime < Date.now()){
+                return (fetchWithoutCache(url, { signal: signal }) as Promise<
+                IfetchResponse<TosuWorldIdsData>>);
+            };
+            return loadUserIdsFromCache(slicedIds, url);
+
+        });
+        promises.push(promise)
+
+
+      }
+
+      const response = await Promise.all(promises).then((responses) => {
+        const data = responses.map((response) => {
+            const responseData = response.data;
+            if(!responseData) return [];
+            if(isFetchError(responseData)) return [];
+            return responseData;
+        });
+        return {
+          data: data.flat(),
+        };
+      });
+
+  
+      const data = response.data;
+      cacheMultipleUsersData(data, ids, urls);
+
+    if(signal?.aborted) return {};
+
+    return response;
+};
+
+const loadUserIdsFromCache = async (ids: string[], url: string): Promise<
+IfetchResponse<TosuWorldIdsData>> => {
+    const urls = [];
+    for(const id of ids){
+        const url = osuWorldApiBase + "users/" + id;
+        urls.push(url);
+
+    }
+    const responses = await loadMultipleUrlsFromCache(urls) as IfetchResponse<TosuWorldIdData>[] | null;
+    if(!responses){
+        return {
+            error: {
+                code: noData,
+                url: url
+            } 
+        }
+    }
+
+    const data = responses.filter(response => {
+        const responseData = response.data;
+        if(!responseData) return false;
+        if((responseData as IFetchError).error !== undefined) return false;
+        return true;
+    }).map((response) => {
+        return response.data as IosuWorldIdSuccess});
+
+    return {
+        data: data,
+        cache: true
+    }
+};
+
+const isFetchError = (data: TosuWorldIdsData | undefined): data is IFetchError => {
+    return (data as IFetchError).error !== undefined;
+  }
+
+const cacheMultipleUsersData = async (data: TosuWorldIdsData | undefined, ids: string[], urls: string[]) => {
+    if(!data) return;
+    if(isFetchError(data)) return;
+    
+    for(const playerData of data){
+        const url = osuWorldApiBase + "users/" + playerData.id;
+        const cacheResponse = {
+            data: playerData,
+            [expireHeader]: genExpireDate(userDataExpireTime),
+        }
+        saveInCache(url, cacheResponse);
+    }
+    const dataIdsSet = new Set(data.map(playerData => playerData.id.toString()));
+
+
+    for(const id of ids){
+        if(dataIdsSet.has(id)) continue;
+        const url = osuWorldApiBase + "users/" + id;
+        saveInCache(url, {
+            data: {
+                error: unknownUserError
+            } 
+        });
+    }
+
+    for (const url of urls){
+        saveInCache(url, {
+            [expireHeader]: genExpireDate(userDataExpireTime),
+        });
+      }
+}
 
 // Modes: osu, taiko, fruits, mania
 const regionRankingUrl =
-    "https://osuworld.octo.moe/api/{{country-code}}/{{country-region-code}}/top/{{mode}}?page={{page}}";
+osuWorldApiBase + "{{country-code}}/{{country-region-code}}/top/{{mode}}?page={{page}}";
 
 const regionRankingExpire = 7200000; //2 hours